Make metadata accessible
Metadata is stored in the individual archive mountables. Accessing those is possible through an entry_mounts() method.

Additionally, the existing methods mount() and unmount() in Archive::SCS are changed to return the archive instead of the controller object. This offers another way to get a reference to each individual archive.