Avoid PaxHeader entries in dist tarball
CPANTS suggests that the pax extended format, which was specified for tar by POSIX.1-2001, is unsupported on some Linux systems. This change is intended to force the use of the even older ustar standard format. However, I'm uncertain if the --format option works with tar variants other than FreeBSD tar and GNU tar.