Lower minimum required version of perl to v5.32
The downsides are many more prereqs, including Feature::Compat::*, builtin::Backport, and Object::Pad. As a result, the installation takes longer than it did before (on a virgin Perl anyway). Also, the boilerplate is slightly longer.

Object::Pad is used directly instead of Feature::Compat::Class because of the feature 'class' bug that prevented using this dist on Perl v5.38.

The main reason I don't want to go lower than Perl v5.32 is the 'isa' feature.

A recent version of Perl (ideally v5.40 or later) is still recommended.