Work around smokers ignoring runtime reqs for test phase
CPAN testers are reporting some weird issues for 1.00, such as  Object::Pad not being installed even though it's a properly declared dependency. According to "Phases" in CPAN::Meta::Spec, runtime requirements must be installed before tests can be run, but apparently these smokers don't do that for some reason. Duplicating the runtime requirements as test requirements might work around that.

This solution is obviously unsatisfactory. It's a stop-gap, and useful to verify if this actually resolves the problem.

Another weird reported issue are perl version requirements. There are several cases of e.g. perl version v5.38 reporting that v5.32 is required, but "this is only v5.16.3". Makes zero sense. Perhaps adding the declaration of the minimum required version of perl to the cpanfile and the use statement (instead of just the runtime reqs) can address this.