Fix cli tests: Use correct perlpath
When adding the tests of the command-line interface, I used just 'perl' for executing the cli script, forgetting that perl may be installed with different names and paths. This change should fix that. To be safe, it also adds a simple check to verify that the version numbers of the cli and the module match up.

Such a setup with perls that aren't named "perl" is rare in regular usage, but it's very common in smoke testing environments.