Make self-contained CityHash build more reliable
CityHash failed to build on compilers that define the SSE4.2 macro if either citycrc.h or _mm_crc32_u64() are unavailable. We could make them both available (the latter via the compiler flag -mcrc32, I think). But we'll never need these functions anyway, so it's probably more reliable to disable them entirely.

Undefining __SSE4_2__ isn't an option because it's predefined by the compiler based on the architecture.