Replace file/handles_file with path/handles_path
The choice of wording for $file, file() and handles_file() was unfortunate: It prevents colliding with Path::Tiny imports, but it also precludes dealing with archives that aren't files (or at least, makes that a bit of a hack, as for InMemory).

This change replaces the old wording with $path, path() and handles_path(). It adds compatibility fallbacks to Mountable to support third-party format handlers that rely on the old methods (although I don't expect there are any, so I'll probably remove these fallbacks soon-ish).