Raise minimum Perl version to v5.34
This change is a work-around for the isa infix bug present in Perl v5.32 (issues 18679 / 20354). This bug actually wasn't fully resolved until v5.36.1, but v5.34 looks good enough for us.

For what it's worth, I'm starting to question whether isa infix is worth it. Yes, it's cleaner and faster, but it might not be that big of a difference. Allowing this code to run on v5.28, which dropping isa infix would do, might be more of an advantage.