Change internal hash format to 16-byte hex string
A hex string is easier to work with in practice than the binary. This should primarily affect debugging.

This change doubles the length of the internal hash string, which means that Perl's own hash function needs to do more work. cityhash64() is 25% slower now. But overall speed is entirely dominated by I/O, so the difference isn't measurable in real-world usage.

This change removes the big-endian special case. The endianness should still be correct, but I can't be sure that CityHash itself even works on big endian. That's untested according to city.h. And I actually don't believe ATS runs on big endian hardware, either.

The hash function is currently not Unicode-aware and croaks on wide characters. Unicode pathnames aren't used in vanilla ATS anyway.